requirements.txt
django
mysqlclient
django-cors-headers
djangorestframework
django-simpleui
Pillow